backend-django/pyproject.toml
[tool.poetry]
name = "superapp-backend-django"
version = "0.1.0"
description = "Backend service for Superapp using Django and Strawberry GraphQL"
authors = ["Ade Musytahun Wahid <ademusytahunwahid@gmail.com>"]
readme = "README.md"  
packages = [{ include = "backend" }]

[tool.poetry.dependencies]
python = "^3.11"
django = ">=5.2.6,<6.0.0"
strawberry-graphql-django = ">=0.65.1,<0.66.0"
django-choices-field = ">=3.1.0,<4.0.0"
django-cors-headers = ">=4.9.0,<5.0.0"
django-environ = ">=0.12.0,<0.13.0"
invoke = ">=2.2.0,<3.0.0"
psycopg2-binary = "^2.9.10"
pyjwt = "^2.10.1"
python-dotenv = "^1.2.1"
dj-database-url = "^3.0.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"